% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get catalog of available data}
\usage{
get_catalog(
  dataset = c("full", "ene", "epf_personas", "epf_gastos", "enusc", "esi")
)
}
\arguments{
\item{dataset}{\code{string} name of dataset. The possible values are "full", "ene", "epf_personas", "epf_gastos", "enusc" or "esi". The default value is "full"}
}
\value{
\code{dataframe}
}
\description{
Get catalog of available data
}
