% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_columns}
\alias{get_columns}
\title{Get catalog of data available}
\usage{
get_columns(dataset, version)
}
\arguments{
\item{dataset}{\code{string}}

\item{version}{\code{string}}
}
\value{
\code{list}
}
\description{
Get catalog of data available
}
