% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download a dataset from the API INE service}
\usage{
get_data(dataset, version = NULL)
}
\arguments{
\item{dataset}{\code{string}. The possible values are "ene", "epf_personas", "epf_gastos", "enusc" or "esi"}

\item{version}{\code{string} by default is the newest version}
}
\value{
\code{dataframe}
}
\description{
Download a dataset from the API INE service
}
