% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get catalog of available data}
\usage{
get_catalog(dataset = c("full", "ene", "epf_personas", "enusc"))
}
\arguments{
\item{dataset}{\code{string} name of dataset}
}
\value{
\code{dataframe}
}
\description{
Get catalog of available data
}
