% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data}
\alias{get_data}
\title{Get catalog of data available}
\usage{
get_data(dataset, version)
}
\arguments{
\item{dataset}{\code{string}}

\item{version}{\code{string}}
}
\value{
\code{dataframe}
}
\description{
Get catalog of data available
}
