% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download a dataset from the API INE service}
\usage{
get_data(
  dataset,
  version = NULL,
  col_list = NULL,
  save_where = c("renviron", "disk", "both", "no_message")
)
}
\arguments{
\item{dataset}{\code{string}. The possible values are "ene", "epf_personas", "epf_gastos", "enusc" or "esi".}

\item{version}{\code{string} by default is the newest version.}

\item{col_list}{\code{character vector} by default is all columns. The possible values can be obtained through the get_columns function.}

\item{save_where}{\code{string} by default is "renviron", the possible values are "renviron","disk","both","no_message".}
}
\value{
\code{dataframe}
}
\description{
Download a dataset from the API INE service
}
\examples{
# Get dataset of "ene" survey from september-november-october 2020 quarter and selecting columns

get_data(dataset="ene", version="2022-10-son", col_list=c("region","cae_general","fact_cal"))

}
