% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_many_data}
\alias{get_many_data}
\title{Get multiple datasets}
\usage{
get_many_data(dataset, from = NULL, to = NULL, versions = NULL)
}
\arguments{
\item{dataset}{\code{string}}

\item{from}{\code{string} specific version of any survey}

\item{to}{\code{string} specific version of any survey}

\item{versions}{\code{character vector} with specific versions of any survey}
}
\value{
\code{list} containing all the datasets between the from and to parameters
}
\description{
Get multiple datasets
}
